Alpha-SQL-master/requirements.txt
datasketch
nltk
chardet
python-dotenv
openai
pandas
loguru
tqdm
pyyaml
weave
sqlglot[rs]